+++ instrumentation/opentelemetry-instrumentation-psycopg2/tests/test_psycopg2_integration.py
@@ -21,6 +21,9 @@ import opentelemetry.instrumentation.psycopg2
 from opentelemetry import trace
 from opentelemetry.instrumentation.psycopg2 import Psycopg2Instrumentor
 from opentelemetry.sdk import resources
+from opentelemetry.semconv._incubating.attributes.db_attributes import (
+    DB_STATEMENT,
+)
 from opentelemetry.test.test_base import TestBase
 
 
@@ -65,6 +68,7 @@ class MockConnection:
         return {"dbname": "test"}
 
 
+# pylint: disable=too-many-public-methods
 class TestPostgresqlIntegration(TestBase):
     def setUp(self):
         super().setUp()
@@ -216,6 +220,96 @@ class TestPostgresqlIntegration(TestBase):
         self.assertEqual(len(spans_list), 1)
 
     # pylint: disable=unused-argument
+    def test_instrument_connection_is_idempotent(self):
+        cnx = psycopg2.connect(database="test")
+        query = "SELECT * FROM test"
+        cursor = cnx.cursor()
+        cursor.execute(query)
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 0)
+
+        instrumentor = Psycopg2Instrumentor()
+        cnx = instrumentor.instrument_connection(cnx)
+        cnx = instrumentor.instrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+    def test_instrument_connection_with_custom_cursor_factory_instrument_then_uninstrument(
+        self,
+    ):
+        instrumentor = Psycopg2Instrumentor()
+        cnx = psycopg2.connect(database="test", cursor_factory=MockCursor)
+        query = "SELECT * FROM test"
+
+        self.assertIs(cnx.cursor_factory, MockCursor)
+
+        cnx = instrumentor.instrument_connection(cnx)
+        self.assertIsNot(cnx.cursor_factory, MockCursor)
+
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+        cnx = instrumentor.uninstrument_connection(cnx)
+        self.assertIs(cnx.cursor_factory, MockCursor)
+
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+    def test_uninstrument_connection_is_idempotent(self):
+        instrumentor = Psycopg2Instrumentor()
+        cnx = psycopg2.connect(database="test")
+        query = "SELECT * FROM test"
+
+        cnx = instrumentor.instrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+        cnx = instrumentor.uninstrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+        cnx = instrumentor.uninstrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+    def test_instrument_connection_reinstrument_after_uninstrument(self):
+        instrumentor = Psycopg2Instrumentor()
+        cnx = psycopg2.connect(database="test")
+        query = "SELECT * FROM test"
+
+        cnx = instrumentor.instrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+        cnx = instrumentor.uninstrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 1)
+
+        cnx = instrumentor.instrument_connection(cnx)
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        spans_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans_list), 2)
+
+    # pylint: disable=unused-argument
     def test_uninstrument_connection_with_instrument(self):
         Psycopg2Instrumentor().instrument()
         cnx = psycopg2.connect(database="test")
@@ -261,6 +355,140 @@ class TestPostgresqlIntegration(TestBase):
         kwargs = event_mocked.call_args[1]
         self.assertEqual(kwargs["enable_commenter"], True)
 
+    def test_sqlcommenter_enabled_instrument_connection_defaults(self):
+        with (
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.__version__",
+                "foobar",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.__libpq_version__",
+                "foobaz",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.threadsafety",
+                "123",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.apilevel",
+                "123",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.paramstyle",
+                "test",
+            ),
+        ):
+            cnx = psycopg2.connect(database="test")
+            cnx = Psycopg2Instrumentor().instrument_connection(
+                cnx,
+                enable_commenter=True,
+            )
+            query = "Select 1"
+            cursor = cnx.cursor()
+            cursor.execute(query)
+            spans_list = self.memory_exporter.get_finished_spans()
+            span = spans_list[0]
+            span_id = format(span.get_span_context().span_id, "016x")
+            trace_id = format(span.get_span_context().trace_id, "032x")
+            trace_flags = int(span.get_span_context().trace_flags)
+            self.assertEqual(
+                MockCursor.execute.call_args[0][0],
+                f"Select 1 /*db_driver='psycopg2%%3Afoobar',dbapi_level='123',dbapi_threadsafety='123',driver_paramstyle='test',libpq_version='foobaz',traceparent='00-{trace_id}-{span_id}-{trace_flags:02x}'*/",
+            )
+            self.assertEqual(
+                span.attributes[DB_STATEMENT],
+                "Select 1",
+            )
+
+    def test_sqlcommenter_enabled_instrument_connection_stmt_enabled(self):
+        with (
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.__version__",
+                "foobar",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.__libpq_version__",
+                "foobaz",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.threadsafety",
+                "123",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.apilevel",
+                "123",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.paramstyle",
+                "test",
+            ),
+        ):
+            cnx = psycopg2.connect(database="test")
+            cnx = Psycopg2Instrumentor().instrument_connection(
+                cnx,
+                enable_commenter=True,
+                enable_attribute_commenter=True,
+            )
+            query = "Select 1"
+            cursor = cnx.cursor()
+            cursor.execute(query)
+            spans_list = self.memory_exporter.get_finished_spans()
+            span = spans_list[0]
+            span_id = format(span.get_span_context().span_id, "016x")
+            trace_id = format(span.get_span_context().trace_id, "032x")
+            trace_flags = int(span.get_span_context().trace_flags)
+            self.assertEqual(
+                MockCursor.execute.call_args[0][0],
+                f"Select 1 /*db_driver='psycopg2%%3Afoobar',dbapi_level='123',dbapi_threadsafety='123',driver_paramstyle='test',libpq_version='foobaz',traceparent='00-{trace_id}-{span_id}-{trace_flags:02x}'*/",
+            )
+            self.assertEqual(
+                span.attributes[DB_STATEMENT],
+                f"Select 1 /*db_driver='psycopg2%%3Afoobar',dbapi_level='123',dbapi_threadsafety='123',driver_paramstyle='test',libpq_version='foobaz',traceparent='00-{trace_id}-{span_id}-{trace_flags:02x}'*/",
+            )
+
+    def test_sqlcommenter_enabled_instrument_connection_with_options(self):
+        with (
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.__version__",
+                "foobar",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.__libpq_version__",
+                "foobaz",
+            ),
+            mock.patch(
+                "opentelemetry.instrumentation.psycopg2.psycopg2.threadsafety",
+                "123",
+            ),
+        ):
+            cnx = psycopg2.connect(database="test")
+            cnx = Psycopg2Instrumentor().instrument_connection(
+                cnx,
+                enable_commenter=True,
+                commenter_options={
+                    "dbapi_level": False,
+                    "dbapi_threadsafety": True,
+                    "driver_paramstyle": False,
+                    "foo": "ignored",
+                },
+            )
+            query = "Select 1"
+            cursor = cnx.cursor()
+            cursor.execute(query)
+            spans_list = self.memory_exporter.get_finished_spans()
+            span = spans_list[0]
+            span_id = format(span.get_span_context().span_id, "016x")
+            trace_id = format(span.get_span_context().trace_id, "032x")
+            trace_flags = int(span.get_span_context().trace_flags)
+            self.assertEqual(
+                MockCursor.execute.call_args[0][0],
+                f"Select 1 /*db_driver='psycopg2%%3Afoobar',dbapi_threadsafety='123',libpq_version='foobaz',traceparent='00-{trace_id}-{span_id}-{trace_flags:02x}'*/",
+            )
+            self.assertEqual(
+                span.attributes[DB_STATEMENT],
+                "Select 1",
+            )
+
     @mock.patch("opentelemetry.instrumentation.dbapi.wrap_connect")
     def test_sqlcommenter_disabled(self, event_mocked):
         cnx = psycopg2.connect(database="test")
@@ -270,6 +498,45 @@ class TestPostgresqlIntegration(TestBase):
         cursor.execute(query)
         kwargs = event_mocked.call_args[1]
         self.assertEqual(kwargs["enable_commenter"], False)
+
+    def test_sqlcommenter_disabled_default_instrument_connection(self):
+        cnx = psycopg2.connect(database="test")
+        cnx = Psycopg2Instrumentor().instrument_connection(
+            cnx,
+        )
+        query = "Select 1"
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        self.assertEqual(
+            MockCursor.execute.call_args[0][0],
+            "Select 1",
+        )
+        spans_list = self.memory_exporter.get_finished_spans()
+        span = spans_list[0]
+        self.assertEqual(
+            span.attributes[DB_STATEMENT],
+            "Select 1",
+        )
+
+    def test_sqlcommenter_disabled_explicit_instrument_connection(self):
+        cnx = psycopg2.connect(database="test")
+        cnx = Psycopg2Instrumentor().instrument_connection(
+            cnx,
+            enable_commenter=False,
+        )
+        query = "Select 1"
+        cursor = cnx.cursor()
+        cursor.execute(query)
+        self.assertEqual(
+            MockCursor.execute.call_args[0][0],
+            "Select 1",
+        )
+        spans_list = self.memory_exporter.get_finished_spans()
+        span = spans_list[0]
+        self.assertEqual(
+            span.attributes[DB_STATEMENT],
+            "Select 1",
+        )
 
     def test_no_op_tracer_provider(self):
         Psycopg2Instrumentor().instrument(